break down test code
